net: hns3: clear hardware resource when loading driver
driver inclusion
category: bugfix
bugzilla: NA
CVE: NA

----------------------------

If a PF is bonded to a virtual machine and the virtual machine
exits unexpectedly, some hardware resource cannot be cleared.
In this case, loading driver may cause exceptions. Therefore,
the hardware resource needs to be cleared when the driver is loaded.

Signed-off-by: Yufeng Mo <moyufeng@huawei.com>
Reviewed-by: Jian Shen <shenjian15@huawei.com>
Reviewed-by: Peng Li <lipeng321@huawei.com>
Reviewed-by: Fenglong Kang <kangfenglong@huawei.com>
Reviewed-by: Liubin Shu <shuliubin@huawei.com>
Signed-off-by: Yonglong Liu <liuyonglong@huawei.com>
Reviewed-by: Yongxin Li <liyongxin1@huawei.com>
Signed-off-by: Yang Yingliang <yangyingliang@huawei.com>