modpost: use read_text_file() and get_line() for reading text files
mainline inclusion
from mainline-5.8
commit 70f30cfe5b892fcb7f98e7df72ed6ccfe3225628
category: bugfix
bugzilla: 46847
CVE: NA

------------------------------------------------

grab_file() mmaps a file, but it is not so efficient here because
get_next_line() copies every line to the temporary buffer anyway.

read_text_file() and get_line() are simpler. get_line() exploits the
library function strchr().

Going forward, the missing *.symvers or *.cmd is a fatal error.
This should not happen because scripts/Makefile.modpost guards the
-i option files with $(wildcard $(input-symdump)).

Signed-off-by: Masahiro Yamada <masahiroy@kernel.org>

Drop the changes in modpost.c since the missing *.symvers or *.cmd is
a fatal error and we don't check it in scripts/Makefile.modpost
currently, on which curimistance the building process for modules would
be interrupted with an error. This isn't what we expect.

Signed-off-by: Wang Wensheng <wangwensheng4@huawei.com>
Reviewed-by: Jian Cheng <cj.chengjian@huawei.com>
Signed-off-by: Yang Yingliang <yangyingliang@huawei.com>