arm64/mpam: Add mpam driver discovery phase and kbuild boiler plate
hulk inclusion
category: feature
bugzilla: 34278
CVE: NA

-------------------------------------------------

Components with MPAM controls (or monitors) could be placed anywhere
in the system, each with their own set MMIO configuration area.

Firmware tables tell us the components position in the topology and
location of its MMIO configuration area, as well as which logical
component this corresponds with.

For now, we are only interested in the well-known caches, e.g. L2.

To reduce the number of times we have to schedule work on another CPU,
we collect all the information from the firmware tables between
mpam_discovery_start() and mpam_discovery_complete().

If the code parseing firmware tables concludes it can't continue, it
can call mpam_discovery_failed() to free the allocated memory.

[Wang ShaoBo: few version adaptation changes]

Signed-off-by: James Morse <james.morse@arm.com>
Link: http://www.linux-arm.org/git?p=linux-jm.git;a=patch;h=ba79c6b4021fb13f395ea033bf950894f1c494fd
Signed-off-by: Wang ShaoBo <bobo.shaobowang@huawei.com>
Reviewed-by: Xiongfeng Wang <wangxiongfeng2@huawei.com>
Reviewed-by: Cheng Jian <cj.chengjian@huawei.com>
Signed-off-by: Yang Yingliang <yangyingliang@huawei.com>