arm64/mpam: Summarize feature support during mpam_enable()
hulk inclusion
category: feature
bugzilla: 34278
CVE: NA

-------------------------------------------------

To make a decision about whether to expose an mpam class as
a resctrl resource we need to know its overall supported
features and properties.

Once we've probed all the devices, we can walk the tree
and produced overall values. If bitmap properties are mismatched
within a component we cannot support that bitmap.

[Wang ShaoBo: few version adaption changes]

Signed-off-by: James Morse <james.morse@arm.com>
Link: http://www.linux-arm.org/git?p=linux-jm.git;a=patch;h=96d7c0b933c0334abf9c45d73f239ac3f382c074
Signed-off-by: Wang ShaoBo <bobo.shaobowang@huawei.com>
Reviewed-by: Xiongfeng Wang <wangxiongfeng2@huawei.com>
Reviewed-by: Cheng Jian <cj.chengjian@huawei.com>
Signed-off-by: Yang Yingliang <yangyingliang@huawei.com>