arm64/mpam: resctrl: Write and read schemata by schema_list
hulk inclusion
category: feature
bugzilla: 34278
CVE: NA

-------------------------------------------------

schemata labels each row with hw_closid, which can be parsed into closid
according to fixed rules (LxCODE and MBA are given closid, LxDATA is
given to closid + 1), so the maximum number of rdtgroup can be created
is also restricted by half if cdp enabled.

The length of Lx Cache domains displayed in schemata is compressed, this
is because for specified hardware platform, domains of each resource may
be too many to be easily operated for user interaction.

This patch also move parse_cbm() and parse_bw() to mpam_resctrl.c for clarity.

Signed-off-by: Wang ShaoBo <bobo.shaobowang@huawei.com>
Reviewed-by: Xiongfeng Wang <wangxiongfeng2@huawei.com>
Reviewed-by: Cheng Jian <cj.chengjian@huawei.com>
Signed-off-by: Yang Yingliang <yangyingliang@huawei.com>