arm64/mpam: Support cdp on allocating monitors
hulk inclusion
category: feature
bugzilla: 34278
CVE: NA

-------------------------------------------------

This prepares for simultaneously monitoring LxDATA and LxCODE when cdp
is open, under our implementation, LxDATA and LxCODE is allocated closid
and closid+1, so we should keep two monitor once time for each.

Why there needs one monitors for each closid when cdp is open, but not
switch one between the two LxDATA and LxCODE is because this monitor kept
by target closid maybe busy for a long time, it would cause inaccuracy if
we force switching.

Signed-off-by: Wang ShaoBo <bobo.shaobowang@huawei.com>
Reviewed-by: Xiongfeng Wang <wangxiongfeng2@huawei.com>
Reviewed-by: Cheng Jian <cj.chengjian@huawei.com>
Signed-off-by: Yang Yingliang <yangyingliang@huawei.com>