arm64/mpam: resctrl: Remove ctrlmon sysfile
hulk inclusion
category: feature
bugzilla: 34278
CVE: NA

-------------------------------------------------

We redesign monitoring process for user, as following illustrates:

  e.g. before rewriting:
     mount /sys/fs/resctrl && cd /sys/fs/resctrl
	 mkdir p1 && cd p1
     echo 1 > ctrlmon  # this allocates a monitor resource for this group
     ...               # associating task/cpu with this group
     grep . mon_data/* # get monitor data from mon_data directory
  e.g. after rewriting:
	 mount /sys/fs/resctrl && cd /sys/fs/resctrl
 	 mkdir p1 && cd p1 # automically allocating a monitoring resource
     ...               # associate task/cpu with this group
     grep . mon_data/* # directly get monitor data

ctrlmon is used for manually allocating a monitor resource for monitoring
a specified group (labeled by partid and pmg), we delete ctrlmon because
this action is redundant.

User should know which group has been allocated a available monitor resource
and only this monitor resource is released then this monitor resource can
be reallocated to a new group after, this action is redundant and unnecessary,
as monitor resource is used only when monitoring process happens, so a
relax monitor resource can be allocated to multiple groups and take effect
when monitoring process happened.

But should some restrictions be known, a monitor resource for monitoring
Cache-occupancy might be kept for a long time until it doesn't need to
be use anymore, or below a threshold as like intel-RDT limbo list works,
otherwise you may see that the monitoring result is very small beyond
exception when you force switch one mon resource from one group to another.

We deliver a simple LRU mon resource allocation mechanism, but so far
it just assign a monitor according to the order in which groups was created,
this is incomplete and needs subsequent improvement.

Signed-off-by: Wang ShaoBo <bobo.shaobowang@huawei.com>
Reviewed-by: Xiongfeng Wang <wangxiongfeng2@huawei.com>
Reviewed-by: Cheng Jian <cj.chengjian@huawei.com>
Signed-off-by: Yang Yingliang <yangyingliang@huawei.com>