share_pool: Calculate k2u size for a task
ascend inclusion
category: feature
bugzilla: 46867
CVE: NA

-------------------------------------------------
Add a new column SP_K2U in /proc/sharepool/proc_stat which means k2u size
for a task.
If a task is the target of k2u(to task) but never added to a sp group, its
sp group id is notated as '-'.

Signed-off-by: Tang Yizhou <tangyizhou@huawei.com>
Reviewed-by: Ding Tianhong <dingtianhong@huawei.com>
Signed-off-by: Yang Yingliang <yangyingliang@huawei.com>