share_pool: Fix a potential bug branch
ascend inclusion
category: bugfix
bugzilla: 46885
CVE: NA

-------------------

When we add a task to a new group, we create and initialize the group at
first. The is_active element of spg should not be true until the fully
process completed, or the BUG of freeing an active spg shall occur if the
works after the initialization raise an error.

Signed-off-by: Wang Wensheng <wangwensheng4@huawei.com>
Reviewed-by: Ding Tianhong <dingtianhong@huawei.com>
Reviewed-by: Tang Yizhou <tangyizhou@huawei.com>
Signed-off-by: Yang Yingliang <yangyingliang@huawei.com>