sharepool: Fix null pointer dereference on adding exiting task
ascend inclusion
category: bugfix
bugzilla: 46885
CVE: NA

-------------------

When we add a task which is exiting to a group, the mm of the task
shall be reset to NULL unexpectedly even if we increase the refcnt
of the task. Then if we access the mm's element via the task_struct,
a NULL pointer dereference issue could occur.

Get and store the mm_struct to be used at beginning to fix the problem.

Signed-off-by: Wang Wensheng <wangwensheng4@huawei.com>
Reviewed-by: Ding Tianhong <dingtianhong@huawei.com>
Reviewed-by: Tang Yizhou <tangyizhou@huawei.com>
Signed-off-by: Yang Yingliang <yangyingliang@huawei.com>