ascend: share pool: optimize the big lock for memory processing
ascend inclusion
category: feature
bugzilla: NA
CVE: NA

-------------------------------------------------

The sp_mutex is used to protect all critical path for share pool,
it has serious affected the performance of the the memory alloc
and release interface when there is a lot of process in the same
memory group, it will serious break the scailability of the system,
so add a new read semaphore lock to instead of the big lock for allocation
and release critical path.

The scailability has been greatly improved by this modification.

Show the test result:
	  		number of process:	alloc 4M avg time:
Before the patch:	1			32us
			3			96us
			10			330us

after the patch:	1			32us
			3			40us
			10			60us

v2: fix some conflicts and clean some code.

Signed-off-by: Ding Tianhong <dingtianhong@huawei.com>
Reviewed-by: Tang Yizhou <tangyizhou@huawei.com>
Reviewed-by: Kefeng Wang <wangkefeng.wang@huawei.com>
Signed-off-by: Yang Yingliang <yangyingliang@huawei.com>