share_pool: Increase refcount of sp_group when call __sp_find_spg
ascend inclusion
category: perf
bugzilla: 47462
CVE: NA

-------------------------------------------------

After getting the pointer of sp_group spg by calling __sp_find_spg(),
the memory of spg may be released if spg is dead and free_sp_group() is
called.

To solve this problem, we increase the refcount of spg when call
__sp_find_spg(). Users should call sp_group_drop() after finishing using
it.

Signed-off-by: Tang Yizhou <tangyizhou@huawei.com>
Reviewed-by: Ding Tianhong <dingtianhong@huawei.com>
Reviewed-by: Kefeng Wang <wangkefeng.wang@huawei.com>
Signed-off-by: Yang Yingliang <yangyingliang@huawei.com>
Signed-off-by: Zheng Zengkai <zhengzengkai@huawei.com>