block_dump: remove block_dump feature when dirting inode
hulk inclusion
category: bugfix
bugzilla: 48166
CVE: NA
---------------------------

block_dump is an old debugging interface, one of it's functions is used
to dump who write which file on disk. If block_dump is enabled, we can
turn on debug log level and gather information about write process name
file name from kmsg. It is done by block_dump___mark_inode_dirty() to
print kernel message directly when marking inode dirty, so it can
trigger log storm easily.

After tracepoints has been introduced into the kernel, we got
trace_writeback_mark_inode_dirty() in __mark_inode_dirty(), which is a
better replacement of block_dump___mark_inode_dirty(). The only downside
is that it only trace the inode number and not a file name, but it may
not a big deal because the original dumped file name in block_dump is
not accurate in some cases, and we can still find it through the inode
number and device id. So this patch delete the block_dump feature.

Signed-off-by: zhangyi (F) <yi.zhang@huawei.com>
Reviewed-by: Ye bin <yebin10@huawei.com>
Signed-off-by: Yang Yingliang <yangyingliang@huawei.com>
Signed-off-by: Zheng Zengkai <zhengzengkai@huawei.com>