iommu/io-pgtable: Replace IO_PGTABLE_QUIRK_NO_DMA with specific flag
mainline inclusion
from mainline-v5.3-rc1
commit 4f41845b340783eaec9cc2840fe3cb9a00574054
category: bugfix
bugzilla: 21306
CVE: NA

-------------------------------------------------------------------------

IO_PGTABLE_QUIRK_NO_DMA is a bit of a misnomer, since it's really just
an indication of whether or not the page-table walker for the IOMMU is
coherent with the CPU caches. Since cache coherency is more than just a
quirk, replace the flag with its own field in the io_pgtable_cfg
structure.

Cc: Bjorn Andersson <bjorn.andersson@linaro.org>
Signed-off-by: Will Deacon <will@kernel.org>
Signed-off-by: Zhen Lei <thunder.leizhen@huawei.com>
Reviewed-by: Hanjun Guo <guohanjun@huawei.com>
Signed-off-by: Yang Yingliang <yangyingliang@huawei.com>