iommu/arm-smmu-v3: Increase maximum size of queues
mainline inclusion
from mainline-v5.3-rc1
commit d25f6ead162eab3f51b6616be23691ac42e141b5
category: bugfix
bugzilla: 21306
CVE: NA

-------------------------------------------------------------------------

We've been artificially limiting the size of our queues to 4k so that we
don't end up allocating huge amounts of physically-contiguous memory at
probe time. However, 4k is only enough for 256 commands in the command
queue, so instead let's try to allocate the largest queue that the SMMU
supports, retrying with a smaller size if the allocation fails.

The caveat here is that we have to limit our upper bound based on
CONFIG_CMA_ALIGNMENT to ensure that our queue allocations remain
natually aligned, which is required by the SMMU architecture.

Signed-off-by: Will Deacon <will.deacon@arm.com>
Conflicts:
	drivers/iommu/arm-smmu-v3.c

Signed-off-by: Zhen Lei <thunder.leizhen@huawei.com>
Reviewed-by: Hanjun Guo <guohanjun@huawei.com>
Signed-off-by: Yang Yingliang <yangyingliang@huawei.com>