iommu/arm-smmu-v3: Reduce contention during command-queue insertion
mainline inclusion
from mainline-v5.4-rc1
commit 587e6c10a7ce89a5924fdbeff2ec524fbd6a124b
category: bugfix
bugzilla: 21306
CVE: NA

-------------------------------------------------------------------------

The SMMU command queue is a bottleneck in large systems, thanks to the
spin_lock which serialises accesses from all CPUs to the single queue
supported by the hardware.

Attempt to improve this situation by moving to a new algorithm for
inserting commands into the queue, which is lock-free on the fast-path.

Tested-by: Ganapatrao Kulkarni  <gkulkarni@marvell.com>
Signed-off-by: Will Deacon <will@kernel.org>
Conflicts:
	drivers/iommu/arm-smmu-v3.c

Signed-off-by: Zhen Lei <thunder.leizhen@huawei.com>
Reviewed-by: Hanjun Guo <guohanjun@huawei.com>
Signed-off-by: Yang Yingliang <yangyingliang@huawei.com>