arm64: Add config switch and kernel parameter for CPU0 hotplug
hulk inclusion
category: feature
bugzilla: 51775
CVE: NA

-------------------------------------------------

New config switch CONFIG_ARM64_BOOTPARAM_HOTPLUG_CPU0 sets default
state of whether the CPU0 hotplug is on or off.

If the config switch is off, CPU0 is not hotpluggable by default.
But the CPU0 hotplug feature can still be turned on by kernel
parameter arm64_cpu0_hotplug at boot.

If the config switch is on, CPU0 is always hotpluggable.

The default value of the config switch is off.

Signed-off-by: Zheng Zengkai <zhengzengkai@huawei.com>
Reviewed-by: Hanjun Guo <guohanjun@huawei.com>
Signed-off-by: Yang Yingliang <yangyingliang@huawei.com>