ipmi: Fix return value when a message is truncated
mainline inclusion
from mainline-5.1
commit 95ac0daa0ca98ad8ac23769021144f8bf04cf406
category: bugfix
bugzilla: NA
CVE: NA

-------------------------------------------------

The user can ask the message to be returned even if it didn't supply
enough memory for the data, and it will return an error but still
fills in as much data as possible.   However, the return value
wasn't being set to an error, it was being overwritten.  Create a
second return value for that case.

Signed-off-by: Corey Minyard <cminyard@mvista.com>
Signed-off-by: Wenchao Hao <haowenchao@huawei.com>
Reviewed-by: Miaohe Lin <linmiaohe@huawei.com>
Signed-off-by: Yang Yingliang <yangyingliang@huawei.com>