irqchip/git-v3-its: Implement irq_retrigger callback for device-triggered LPIs
mainline inclusion
from mainline-5.10
commit 5f774f5e12512b850a611aa99b4601d7eac50edb
category: bugfix
bugzilla: NA
CVE: NA

-------------------------------------------------

It is pretty easy to provide a retrigger callback for the ITS,
as it we already have the required support in terms of
irq_set_irqchip_state().

Note that this only works for device-generated LPIs, and not
the GICv4 doorbells, which should never have to be retriggered
anyway.

Reviewed-by: Valentin Schneider <valentin.schneider@arm.com>
Signed-off-by: Marc Zyngier <maz@kernel.org>
Signed-off-by: Liao Chang <liaochang1@huawei.com>
Reviewed-by: Hanjun Guo <guohanjun@huawei.com>
Signed-off-by: Yang Yingliang <yangyingliang@huawei.com>