net: hns3: modify an error type configuration
driver inclusion
category: bugfix
bugzilla: NA
CVE: NA

----------------------------

According to the UM, the error configuration of the igu_egu_hw_err
is that the type and enable status should be configured separately.
The error type is also changed by mistake when we disable it
currently. The correct way is to configure these two parameters
separately.

Signed-off-by: Yufeng Mo <moyufeng@huawei.com>
Signed-off-by: Yonglong Liu <liuyonglong@huawei.com>
Reviewed-by: li yongxin <liyongxin1@huawei.com>
Signed-off-by: Yang Yingliang <yangyingliang@huawei.com>